Moved OAuth functionality to its own classes.
This is no where near finished, it needs some major polish, refinement, refactoring, and just a few happy thoughts.

The main goal of this commit was to get the OAuth code out of the core plugin class, I will be finishing and polishing it over the next few, and possibly last, commits of branch 1.1.0 as 1.1.0 was meant to be the branch dedicated to the development of the OAuth functionality.